[#677] Account secrets (#21)
* start implementing account secrets - added to core and API models, added encryption, decryption utils [#677]

* implement service AES key management [#677]

* update mock storage [#677]

* add update secrets API, add decrypt secrets func to auth interface [#677]

* add missing files

* implement secrets decryption on get account and login [#677]

* bug fixes and cleanup, add timestamp fields to key model (RSA-PSS parsing not working) [#677]

* update changelog
