+++ .secrets.baseline
@@ -76,10 +76,6 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
-      "path": "detect_secrets.filters.common.is_baseline_file",
-      "filename": ".secrets.baseline"
-    },
-    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -137,52 +133,73 @@
         "hashed_secret": "44e17306b837162269a410204daaa5ecee4ec22c",
         "is_verified": false,
         "line_number": 42
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "core/app_shared.go",
+        "hashed_secret": "48da876cb7879f4d7e8d5e679b8fbd0af02845ae",
+        "is_verified": false,
+        "line_number": 80
       }
     ],
     "core/auth/apis.go": [
       {
         "type": "Secret Keyword",
         "filename": "core/auth/apis.go",
+        "hashed_secret": "48da876cb7879f4d7e8d5e679b8fbd0af02845ae",
+        "is_verified": false,
+        "line_number": 167
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "core/auth/apis.go",
         "hashed_secret": "04e110541a2e8b44bc10939bfaf5d82adfe45158",
         "is_verified": false,
-        "line_number": 1944
+        "line_number": 1952
       },
       {
         "type": "Secret Keyword",
         "filename": "core/auth/apis.go",
         "hashed_secret": "394e3412459f79523e12e1fa95a4cf141ccff122",
         "is_verified": false,
-        "line_number": 2277
+        "line_number": 2347
       }
     ],
     "core/auth/auth.go": [
       {
         "type": "Secret Keyword",
         "filename": "core/auth/auth.go",
-        "hashed_secret": "3fea7ef2cdd6ecf5280c66dbd062272664559d83",
+        "hashed_secret": "70fd094756d9fcdcdfbd372dc526e9068a0ee711",
         "is_verified": false,
-        "line_number": 160
+        "line_number": 165
       },
       {
         "type": "Secret Keyword",
         "filename": "core/auth/auth.go",
-        "hashed_secret": "4a0043e461375664a5656fbdda0d3c39a42a1af4",
+        "hashed_secret": "a358987289cd70bbf50fb10acbcb9bff73c66df6",
         "is_verified": false,
-        "line_number": 162
+        "line_number": 167
       },
       {
         "type": "Secret Keyword",
         "filename": "core/auth/auth.go",
         "hashed_secret": "58f3388441fbce0e48aef2bf74413a6f43f6dc70",
         "is_verified": false,
-        "line_number": 982
+        "line_number": 991
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "core/auth/auth.go",
+        "hashed_secret": "44e17306b837162269a410204daaa5ecee4ec22c",
+        "is_verified": false,
+        "line_number": 2676
       },
       {
         "type": "Secret Keyword",
         "filename": "core/auth/auth.go",
         "hashed_secret": "94a7f0195bbbd2260c4e4d02b6348fbcd90b2b30",
         "is_verified": false,
-        "line_number": 2730
+        "line_number": 2806
       }
     ],
     "core/auth/auth_type_oidc.go": [
@@ -246,13 +263,29 @@
         "line_number": 223
       }
     ],
+    "driven/storage/conversions_user.go": [
+      {
+        "type": "Secret Keyword",
+        "filename": "driven/storage/conversions_user.go",
+        "hashed_secret": "6a1a14990e40a6eaa33109699e00c7eddf89334b",
+        "is_verified": false,
+        "line_number": 32
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "driven/storage/conversions_user.go",
+        "hashed_secret": "29a814fd3871dac2897523fce754e9efa56a2467",
+        "is_verified": false,
+        "line_number": 67
+      }
+    ],
     "driven/storage/database.go": [
       {
         "type": "Secret Keyword",
         "filename": "driven/storage/database.go",
         "hashed_secret": "6547f385c6d867e20f8217018a4d468a7d67d638",
         "is_verified": false,
-        "line_number": 224
+        "line_number": 232
       }
     ],
     "driven/storage/migrations.go": [
@@ -280,6 +313,15 @@
         "line_number": 637
       }
     ],
+    "driver/web/conversions_user.go": [
+      {
+        "type": "Secret Keyword",
+        "filename": "driver/web/conversions_user.go",
+        "hashed_secret": "fe86558143c0bd528f649a153bdc32b8fa90301c",
+        "is_verified": false,
+        "line_number": 67
+      }
+    ],
     "driver/web/docs/gen/def.yaml": [
       {
         "type": "Secret Keyword",
@@ -295,7 +337,7 @@
         "filename": "driver/web/docs/gen/gen_types.go",
         "hashed_secret": "c9739eab2dfa093cc0e450bf0ea81a43ae67b581",
         "is_verified": false,
-        "line_number": 1920
+        "line_number": 1921
       }
     ],
     "driver/web/docs/resources/admin/auth/login.yaml": [
@@ -324,7 +366,16 @@
         "is_verified": false,
         "line_number": 24
       }
+    ],
+    "main.go": [
+      {
+        "type": "Secret Keyword",
+        "filename": "main.go",
+        "hashed_secret": "44e17306b837162269a410204daaa5ecee4ec22c",
+        "is_verified": false,
+        "line_number": 136
+      }
     ]
   },
-  "generated_at": "2023-10-03T21:38:39Z"
+  "generated_at": "2023-10-06T23:08:57Z"
 }
