+++ core/app_administration.go
@@ -1072,10 +1072,6 @@ func (app *application) admGetAccounts(limit int, offset int, appID string, orgI
 	return accounts, nil
 }
 
-func (app *application) admGetAccount(accountID string) (*model.Account, error) {
-	return app.getAccount(nil, accountID)
-}
-
 func (app *application) admGetAccountSystemConfigs(appID string, orgID string, accountID string, l *logs.Log) (map[string]interface{}, error) {
 	//find the account
 	account, err := app.getAccount(nil, accountID)
