+++ core/app_services.go
@@ -37,10 +37,6 @@ func (app *application) serGetProfile(accountID string) (*model.Profile, error) 
 	return &profile, nil
 }
 
-func (app *application) serGetAccount(accountID string) (*model.Account, error) {
-	return app.getAccount(nil, accountID)
-}
-
 func (app *application) serGetPreferences(accountID string) (map[string]interface{}, error) {
 	//find the account
 	account, err := app.storage.FindAccountByID(nil, accountID)
@@ -130,6 +126,20 @@ func (app *application) serUpdateAccountPreferences(id string, appID string, org
 		return false, errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccountPreferences, nil, err)
 	}
 	return false, nil
+}
+
+func (app *application) serUpdateAccountSecrets(accountID string, secrets map[string]interface{}) error {
+	encryptedSecrets, err := app.auth.EncryptSecrets(secrets)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionEncrypt, model.TypeAccountSecrets, nil, err)
+	}
+
+	err = app.storage.UpdateAccountSecrets(nil, accountID, encryptedSecrets)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccountSecrets, &logutils.FieldArgs{"id": accountID}, err)
+	}
+
+	return nil
 }
 
 func (app *application) serDeleteAccount(id string) error {
