+++ core/app_shared.go
@@ -67,6 +67,21 @@ func (app *application) sharedGetAppConfig(appTypeIdentifier string, orgID *stri
 	return appConfigs, nil
 }
 
+func (app *application) sharedGetAccount(accountID string) (*model.Account, error) {
+	account, err := app.getAccount(nil, accountID)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionGet, model.TypeAccount, nil, err)
+	}
+
+	decryptedSecrets, err := app.auth.DecryptSecrets(account.Secrets)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionDecrypt, model.TypeAccountSecrets, &logutils.FieldArgs{"id": accountID}, err)
+	}
+	account.Secrets = decryptedSecrets
+
+	return account, nil
+}
+
 func (app *application) sharedGetAccountsByParams(searchParams map[string]interface{}, appID string, orgID string, limit int, offset int, allAccess bool, approvedKeys []string) ([]map[string]interface{}, error) {
 	accounts, err := app.storage.FindAccountsByParams(searchParams, appID, orgID, limit, offset, allAccess, approvedKeys)
 	if err != nil {
