+++ core/auth/apis.go
@@ -159,6 +159,14 @@ func (a *Auth) Login(ipAddress string, deviceType string, deviceOS *string, devi
 	}
 
 	if loginSession.State == "" {
+		if loginSession.Account != nil {
+			decryptedSecrets, err := a.DecryptSecrets(loginSession.Account.Secrets)
+			if err != nil {
+				return nil, nil, nil, errors.WrapErrorAction(logutils.ActionDecrypt, model.TypeAccountSecrets, nil, err)
+			}
+			loginSession.Account.Secrets = decryptedSecrets
+		}
+
 		return nil, loginSession, nil, nil
 	}
 
@@ -2197,6 +2205,68 @@ func (a *Auth) CheckGroups(context storage.TransactionContext, appOrg *model.App
 	}
 
 	return groups, nil
+}
+
+// EncryptSecrets JSON encodes and encrypts the given plain secrets
+func (a *Auth) EncryptSecrets(secrets map[string]interface{}) (map[string]interface{}, error) {
+	if len(secrets) == 0 {
+		return nil, nil
+	}
+
+	secretsStr, err := json.Marshal(secrets)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionMarshal, "secrets", nil, err)
+	}
+
+	data, dataNonce, key, keyNonce, err := utils.Encrypt(secretsStr, a.serviceAESKey)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionEncrypt, "secrets", nil, err)
+	}
+
+	encryptedSecrets := map[string]interface{}{
+		"data":       data,
+		"data_nonce": dataNonce,
+		"key":        key,
+		"key_nonce":  keyNonce,
+	}
+	return encryptedSecrets, nil
+}
+
+// DecryptSecrets decrypts and JSON decodes the given encrypted secrets
+func (a *Auth) DecryptSecrets(secrets map[string]interface{}) (map[string]interface{}, error) {
+	if len(secrets) == 0 {
+		return nil, nil
+	}
+
+	data, ok := secrets["data"].(string)
+	if !ok {
+		return nil, errors.ErrorData(logutils.StatusMissing, "encrypted data", nil)
+	}
+	dataNonce, ok := secrets["data_nonce"].(string)
+	if !ok {
+		return nil, errors.ErrorData(logutils.StatusMissing, "encrypted data nonce", nil)
+	}
+	key, ok := secrets["key"].(string)
+	if !ok {
+		return nil, errors.ErrorData(logutils.StatusMissing, "encrypted key", nil)
+	}
+	keyNonce, ok := secrets["key_nonce"].(string)
+	if !ok {
+		return nil, errors.ErrorData(logutils.StatusMissing, "encrypted key nonce", nil)
+	}
+
+	decrypted, err := utils.Decrypt(data, dataNonce, key, keyNonce, a.serviceAESKey)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionDecrypt, "secrets", nil, err)
+	}
+
+	var decryptedSecrets map[string]interface{}
+	err = json.Unmarshal(decrypted, &decryptedSecrets)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionUnmarshal, "secrets", nil, err)
+	}
+
+	return decryptedSecrets, nil
 }
 
 // GetServiceRegistrations retrieves all service registrations
