+++ driver/web/conversions_user.go
@@ -28,6 +28,8 @@ func accountToDef(item model.Account) *Def.Account {
 	privacy := privacyToDef(&item.Privacy)
 	//preferences
 	preferences := &item.Preferences
+	//secrets
+	secrets := &item.Secrets
 	//systemConfigs
 	systemConfigs := &item.SystemConfigs
 	//permissions
@@ -62,7 +64,7 @@ func accountToDef(item model.Account) *Def.Account {
 	}
 
 	return &Def.Account{Id: &item.ID, Anonymous: &item.Anonymous, System: &item.AppOrg.Organization.System, Permissions: &permissions, Roles: &roles, Groups: &groups,
-		Privacy: privacy, Verified: &item.Verified, Scopes: &scopes, Identifiers: &identifiers, AuthTypes: &authTypes, Profile: profile, Preferences: preferences,
+		Privacy: privacy, Verified: &item.Verified, Scopes: &scopes, Identifiers: &identifiers, AuthTypes: &authTypes, Profile: profile, Preferences: preferences, Secrets: secrets,
 		SystemConfigs: systemConfigs, LastLoginDate: &lastLoginDate, LastAccessTokenDate: &lastAccessTokenDate, MostRecentClientVersion: item.MostRecentClientVersion, Username: username}
 }
