+++ driver/web/docs/gen/def.yaml
@@ -1581,6 +1581,31 @@ paths:
           description: Unauthorized
         '500':
           description: Internal error
+  /services/account/secrets:
+    put:
+      tags:
+        - Services
+      summary: Update account secrets
+      description: |
+        Updates account secrets
+
+        **Auth:** Requires user auth token
+      security:
+        - bearerAuth: []
+      responses:
+        '200':
+          description: Success
+          content:
+            text/plain:
+              schema:
+                type: string
+                example: Success
+        '400':
+          description: Bad request
+        '401':
+          description: Unauthorized
+        '500':
+          description: Internal error
   /services/account/system-configs:
     get:
       tags:
@@ -6778,6 +6803,9 @@ components:
           type: object
           nullable: true
         system_configs:
+          type: object
+          nullable: true
+        secrets:
           type: object
           nullable: true
         anonymous:
