+++ driver/web/docs/resources/admin/application/accounts/system-configs.yaml
@@ -66,16 +66,16 @@ put:
           type: object
     required: true
   responses:
-    '200':
+    200:
       description: Success
       content:
         text/plain:
           schema:
             type: string
             example: Success
-    '400':
+    400:
       description: Bad request
-    '401':
+    401:
       description: Unauthorized
-    '500':
+    500:
       description: Internal error
