+++ driver/web/docs/resources/services/account/preferences.yaml
@@ -16,18 +16,18 @@ put:
           type: object
     required: true
   responses:
-    '200':
+    200:
       description: Success
       content:
         text/plain:
           schema:
             type: string
             example: Success
-    '400':
+    400:
       description: Bad request
-    '401':
+    401:
       description: Unauthorized
-    '500':
+    500:
       description: Internal error
 get:
   tags:
