+++ driver/web/docs/resources/services/account/secrets.yaml
@@ -1,0 +1,24 @@
+put:
+  tags:
+  - Services
+  summary: Update account secrets
+  description: |
+    Updates account secrets
+
+    **Auth:** Requires user auth token
+  security:
+    - bearerAuth: []
+  responses:
+     200:
+      description: Success
+      content:
+        text/plain:
+          schema:
+            type: string
+            example: Success
+     400:
+      description: Bad request
+     401:
+      description: Unauthorized
+     500:
+      description: Internal error
