+++ driver/web/docs/schemas/user/Account.yaml
@@ -17,6 +17,9 @@ properties:
   system_configs:
     type: object
     nullable: true
+  secrets:
+    type: object
+    nullable: true
   anonymous:
     type: boolean
   verified:
