+++ main.go
@@ -132,30 +132,18 @@ func main() {
 		logger.Infof("Error parsing legacy signature support, applying defaults: %v", err)
 		supportLegacySigs = true
 	}
-
-	var authPrivKeyPem string
-	authPrivKeyPemString := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_AUTH_PRIV_KEY", false, true)
-	if authPrivKeyPemString != "" {
-		//make it to be a single line - AWS environemnt variable issue
-		authPrivKeyPem = strings.ReplaceAll(authPrivKeyPemString, `\n`, "\n")
-	} else {
-		authPrivateKeyPath := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_AUTH_PRIV_KEY_PATH", true, false)
-		authPrivKeyPemBytes, err := os.ReadFile(authPrivateKeyPath)
-		if err != nil {
-			logger.Fatalf("Could not find auth priv key file: %v", err)
-		}
-
-		authPrivKeyPem = string(authPrivKeyPemBytes)
+	currentAuthPrivKey := parsePrivKeyFromEnvVar("ROKWIRE_CORE_AUTH_PRIV_KEY", envLoader, supportLegacySigs, logger)
+	if currentAuthPrivKey == nil {
+		logger.Fatalf("Cannot parse the current private key: %v", err)
 	}
 
-	alg := keys.PS256
-	if supportLegacySigs {
-		alg = keys.RS256
-	}
-	authPrivKey, err := keys.NewPrivKey(alg, authPrivKeyPem)
+	oldSupportLegacySigsStr := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_OLD_SUPPORT_LEGACY_SIGNATURES", false, false)
+	oldSupportLegacySigs, err := strconv.ParseBool(oldSupportLegacySigsStr)
 	if err != nil {
-		logger.Fatalf("Failed to parse auth priv key: %v", err)
+		logger.Infof("Error parsing old legacy signature support, applying defaults: %v", err)
+		oldSupportLegacySigs = true
 	}
+	oldAuthPrivKey := parsePrivKeyFromEnvVar("ROKWIRE_CORE_OLD_AUTH_PRIV_KEY", envLoader, oldSupportLegacySigs, logger)
 
 	minTokenExpStr := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_MIN_TOKEN_EXP", false, false)
 	var minTokenExp *int64
@@ -192,7 +180,7 @@ func main() {
 		FirstParty:  true,
 	}
 
-	authImpl, err := auth.NewAuth(serviceID, host, authPrivKey, authService, storageAdapter, emailer, twilioPhoneVerifier, profileBBAdapter,
+	authImpl, err := auth.NewAuth(serviceID, host, currentAuthPrivKey, oldAuthPrivKey, authService, storageAdapter, emailer, twilioPhoneVerifier, profileBBAdapter,
 		minTokenExp, maxTokenExp, supportLegacySigs, Version, logger)
 	if err != nil {
 		logger.Fatalf("Error initializing auth: %v", err)
@@ -241,4 +229,34 @@ func main() {
 	webAdapter := web.NewWebAdapter(env, authImpl.ServiceRegManager, port, coreAPIs, host, corsAllowedOrigins,
 		corsAllowedHeaders, baseServerURL, prodServerURL, testServerURL, devServerURL, logger)
 	webAdapter.Start()
+}
+
+func parsePrivKeyFromEnvVar(envVarName string, envLoader envloader.EnvLoader, supportLegacySigs bool, logger *logs.Logger) *keys.PrivKey {
+	var authPrivKeyPem string
+	authPrivKeyPemString := envLoader.GetAndLogEnvVar(envVarName, false, true)
+	if authPrivKeyPemString != "" {
+		//make it to be a single line - AWS environemnt variable issue
+		authPrivKeyPem = strings.ReplaceAll(authPrivKeyPemString, `\n`, "\n")
+	} else {
+		authPrivateKeyPath := envLoader.GetAndLogEnvVar(envVarName+"_PATH", false, false)
+		if authPrivateKeyPath == "" {
+			return nil
+		}
+
+		authPrivKeyPemBytes, err := os.ReadFile(authPrivateKeyPath)
+		if err != nil {
+			logger.Fatalf("Could not find auth priv key file: %v", err)
+		}
+		authPrivKeyPem = string(authPrivKeyPemBytes)
+	}
+
+	alg := keys.PS256
+	if supportLegacySigs {
+		alg = keys.RS256
+	}
+	authPrivKey, err := keys.NewPrivKey(alg, authPrivKeyPem)
+	if err != nil {
+		logger.Fatalf("Failed to parse auth priv key: %v", err)
+	}
+	return authPrivKey
 }
